requirements.txt
numpy==1.24.2
matplotlib==3.6.2
Pillow==9.4.0
opencv-python==4.6.0
tqdm==4.64.1
vizdoom==1.1.14
torch==1.12.1